Moving Arrangement UI components into a dedicated subfolder
