+++ src/caretogether-pwa/src/Referrals/Arrangements/ArrangementCard.tsx
@@ -18,28 +18,28 @@ import {
   Typography,
 } from '@mui/material';
 import { useState } from 'react';
-import { ArrangementPhase, Arrangement, CombinedFamilyInfo, ChildInvolvement, FunctionRequirement, Permission, ArrangementPolicy } from '../GeneratedClient';
-import { useFamilyLookup, usePersonLookup } from '../Model/DirectoryModel';
-import { PersonName } from '../Families/PersonName';
-import { FamilyName } from '../Families/FamilyName';
+import { ArrangementPhase, Arrangement, CombinedFamilyInfo, ChildInvolvement, FunctionRequirement, Permission, ArrangementPolicy } from '../../GeneratedClient';
+import { useFamilyLookup, usePersonLookup } from '../../Model/DirectoryModel';
+import { PersonName } from '../../Families/PersonName';
+import { FamilyName } from '../../Families/FamilyName';
 import { useRecoilValue } from 'recoil';
-import { policyData } from '../Model/ConfigurationModel';
+import { policyData } from '../../Model/ConfigurationModel';
 import PersonPinCircleIcon from '@mui/icons-material/PersonPinCircle';
 import EventIcon from '@mui/icons-material/Event';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import { TrackChildLocationDialog } from './TrackChildLocationDialog';
-import { MissingArrangementRequirementRow } from "../Requirements/MissingArrangementRequirementRow";
-import { ExemptedRequirementRow } from "../Requirements/ExemptedRequirementRow";
-import { CompletedRequirementRow } from "../Requirements/CompletedRequirementRow";
-import { ArrangementContext, RequirementContext } from "../Requirements/RequirementContext";
+import { MissingArrangementRequirementRow } from "../../Requirements/MissingArrangementRequirementRow";
+import { ExemptedRequirementRow } from "../../Requirements/ExemptedRequirementRow";
+import { CompletedRequirementRow } from "../../Requirements/CompletedRequirementRow";
+import { ArrangementContext, RequirementContext } from "../../Requirements/RequirementContext";
 import { ArrangementPhaseSummary } from './ArrangementPhaseSummary';
 import { ArrangementCardTitle } from './ArrangementCardTitle';
 import { ArrangementFunctionRow } from './ArrangementFunctionRow';
-import { useCollapsed } from '../Hooks/useCollapsed';
+import { useCollapsed } from '../../Hooks/useCollapsed';
 import { ArrangementComments } from './ArrangementComments';
-import { useInlineEditor } from '../Hooks/useInlineEditor';
-import { useReferralsModel } from '../Model/ReferralsModel';
-import { useFamilyIdPermissions } from '../Model/SessionModel';
+import { useInlineEditor } from '../../Hooks/useInlineEditor';
+import { useReferralsModel } from '../../Model/ReferralsModel';
+import { useFamilyIdPermissions } from '../../Model/SessionModel';
 import { format } from 'date-fns';
 import { DatePicker } from '@mui/x-date-pickers';
