+++ src/caretogether-pwa/src/Referrals/Arrangements/ArrangementFunctionRow.tsx
@@ -1,12 +1,12 @@
 import { TableCell, TableRow } from "@mui/material";
 import { useState } from "react";
-import { Arrangement, ArrangementFunction, ArrangementPolicy, FamilyVolunteerAssignment, FunctionRequirement, IndividualVolunteerAssignment, Permission } from "../GeneratedClient";
-import { useFamilyLookup, usePersonLookup } from "../Model/DirectoryModel";
-import { useFamilyIdPermissions } from "../Model/SessionModel";
-import { useDialogHandle } from "../Hooks/useDialogHandle";
-import { FamilyName } from "../Families/FamilyName";
-import { PersonName } from "../Families/PersonName";
-import { IconRow } from "../IconRow";
+import { Arrangement, ArrangementFunction, ArrangementPolicy, FamilyVolunteerAssignment, FunctionRequirement, IndividualVolunteerAssignment, Permission } from "../../GeneratedClient";
+import { useFamilyLookup, usePersonLookup } from "../../Model/DirectoryModel";
+import { useFamilyIdPermissions } from "../../Model/SessionModel";
+import { useDialogHandle } from "../../Hooks/useDialogHandle";
+import { FamilyName } from "../../Families/FamilyName";
+import { PersonName } from "../../Families/PersonName";
+import { IconRow } from "../../IconRow";
 import { AssignArrangementFunctionDialog } from "./AssignArrangementFunctionDialog";
 import { UnassignArrangementFunctionDialog } from "./UnassignArrangementFunctionDialog";
