+++ src/caretogether-pwa/src/Referrals/Arrangements/AssignArrangementFunctionDialog.tsx
@@ -1,12 +1,12 @@
 import { useState } from 'react';
 import { Autocomplete, Button, Dialog, DialogActions, DialogContent, DialogTitle, FormControl, FormControlLabel, FormLabel, Grid, Radio, RadioGroup, TextField } from '@mui/material';
-import { ArrangementPolicy, Arrangement, ArrangementFunction, RoleApprovalStatus, Person, Family, ValueTupleOfPersonAndFamilyAdultRelationshipInfo } from '../GeneratedClient';
+import { ArrangementPolicy, Arrangement, ArrangementFunction, RoleApprovalStatus, Person, Family, ValueTupleOfPersonAndFamilyAdultRelationshipInfo } from '../../GeneratedClient';
 import { useRecoilValue } from 'recoil';
 import { useParams } from 'react-router-dom';
-import { useBackdrop } from '../Hooks/useBackdrop';
-import { DialogHandle } from "../Hooks/useDialogHandle";
-import { useReferralsModel } from '../Model/ReferralsModel';
-import { visibleFamiliesData, usePersonAndFamilyLookup } from '../Model/DirectoryModel';
+import { useBackdrop } from '../../Hooks/useBackdrop';
+import { DialogHandle } from '../../Hooks/useDialogHandle';
+import { useReferralsModel } from '../../Model/ReferralsModel';
+import { visibleFamiliesData, usePersonAndFamilyLookup } from '../../Model/DirectoryModel';
 
 interface AssignArrangementFunctionDialogProps {
   handle: DialogHandle
