+++ src/caretogether-pwa/src/Referrals/Arrangements/CreateArrangementDialog.tsx
@@ -1,12 +1,12 @@
 import { useState } from 'react';
 import { Button, Dialog, DialogActions, DialogContent, DialogTitle, FormControl, FormControlLabel, FormLabel, Grid, Radio, RadioGroup, TextField } from '@mui/material';
-import { CombinedFamilyInfo, ArrangementPolicy, ChildInvolvement } from '../GeneratedClient';
-import { visibleFamiliesData } from '../Model/DirectoryModel';
+import { CombinedFamilyInfo, ArrangementPolicy, ChildInvolvement } from '../../GeneratedClient';
+import { visibleFamiliesData } from '../../Model/DirectoryModel';
 import { DatePicker } from '@mui/x-date-pickers';
 import { useRecoilValue } from 'recoil';
 import { useParams } from 'react-router-dom';
-import { useBackdrop } from '../Hooks/useBackdrop';
-import { useReferralsModel } from '../Model/ReferralsModel';
+import { useBackdrop } from '../../Hooks/useBackdrop';
+import { useReferralsModel } from '../../Model/ReferralsModel';
 
 interface CreateArrangementDialogProps {
   referralId: string,
