+++ src/caretogether-pwa/src/Referrals/Arrangements/DeleteArrangementDialog.tsx
@@ -1,8 +1,8 @@
 import { useParams } from 'react-router';
-import { Arrangement, Person } from '../GeneratedClient';
-import { usePersonLookup } from '../Model/DirectoryModel';
-import { useReferralsModel } from '../Model/ReferralsModel';
-import { UpdateDialog } from '../UpdateDialog';
+import { Arrangement, Person } from '../../GeneratedClient';
+import { usePersonLookup } from '../../Model/DirectoryModel';
+import { useReferralsModel } from '../../Model/ReferralsModel';
+import { UpdateDialog } from '../../UpdateDialog';
 
 interface DeleteArrangementDialogProps {
   referralId: string,
