+++ src/caretogether-pwa/src/Referrals/Arrangements/UnassignArrangementFunctionDialog.tsx
@@ -1,11 +1,11 @@
 import { Button, Dialog, DialogActions, DialogContent, DialogTitle } from '@mui/material';
-import { ArrangementPolicy, Arrangement, ArrangementFunction, Person, FamilyVolunteerAssignment, IndividualVolunteerAssignment } from '../GeneratedClient';
-import { useBackdrop } from '../Hooks/useBackdrop';
-import { DialogHandle } from "../Hooks/useDialogHandle";
-import { useReferralsModel } from '../Model/ReferralsModel';
-import { useFamilyLookup, usePersonLookup } from '../Model/DirectoryModel';
-import { PersonName } from '../Families/PersonName';
-import { FamilyName } from '../Families/FamilyName';
+import { ArrangementPolicy, Arrangement, ArrangementFunction, Person, FamilyVolunteerAssignment, IndividualVolunteerAssignment } from '../../GeneratedClient';
+import { useBackdrop } from '../../Hooks/useBackdrop';
+import { DialogHandle } from '../../Hooks/useDialogHandle';
+import { useReferralsModel } from '../../Model/ReferralsModel';
+import { useFamilyLookup, usePersonLookup } from '../../Model/DirectoryModel';
+import { PersonName } from '../../Families/PersonName';
+import { FamilyName } from '../../Families/FamilyName';
 
 interface UnassignArrangementFunctionDialogProps {
   handle: DialogHandle
