+++ src/caretogether-pwa/src/Referrals/PartneringFamilyScreen.tsx
@@ -11,11 +11,11 @@ import { useState } from 'react';
 import { AddAdultDialog } from '../Families/AddAdultDialog';
 import { AddChildDialog } from '../Families/AddChildDialog';
 import { AddEditNoteDialog } from '../Notes/AddEditNoteDialog';
-import { ArrangementCard } from './ArrangementCard';
+import { ArrangementCard } from './Arrangements/ArrangementCard';
 import { format } from 'date-fns';
 import { UploadFamilyDocumentsDialog } from '../Families/UploadFamilyDocumentsDialog';
 import { policyData } from '../Model/ConfigurationModel';
-import { CreateArrangementDialog } from './CreateArrangementDialog';
+import { CreateArrangementDialog } from './Arrangements/CreateArrangementDialog';
 import { CloseReferralDialog } from './CloseReferralDialog';
 import { OpenNewReferralDialog } from './OpenNewReferralDialog';
 import { FamilyDocuments } from '../Families/FamilyDocuments';
